requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.27
python-jose==3.3.0
python-multipart==0.0.7
passlib==1.7.4
psycopg2-binary==2.9.9
google-generativeai==0.3.2
anthropic==0.18.0
beautifulsoup4==4.12.2
python-docx==1.1.0
PyMuPDF==1.23.8
reportlab==4.0.8
python-dotenv==1.0.0
httpx==0.27.0
bcrypt==4.1.2
pillow==10.2.0
alembic==1.13.1
uuid==1.30
requests>=2.28.0